+++ main/CMakeLists.txt
@@ -14,6 +14,7 @@ idf_component_register(SRCS "swadge2024.c"
                             "display/shapes.c"
                             "display/wsg.c"
                             "menu/menu.c"
+                            "menu/menuRenderer.c"
                             "modes/colorchord/mode_colorchord.c"
                             "modes/demo/demoMode.c"
                             "modes/mainMenu/mainMenu.c"
