+++ main/swadge2024.h
@@ -170,6 +170,7 @@
 #include "shapes.h"
 #include "fill.h"
 #include "menu.h"
+#include "menuRenderer.h"
 
 // Asset loaders
 #include "spiffs_wsg.h"
